requirements.txt
accelerate==0.24.1
aiosqlite==0.19.0
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncio-redis==0.16.0
asyncpg==0.29.0
broadcaster==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
diffusers==0.23.1
dnspython==2.4.2
dramatiq==1.15.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.0
filelock==3.13.1
fsspec==2023.10.0
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.3.2
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
motor==3.3.1
mpmath==1.3.0
mypy==1.6.1
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
prometheus-client==0.19.0
psutil==5.9.6
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pymongo==4.5.0
pytest==7.4.2
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.1
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
websockets==12.0
zipp==3.17.0
